Update code.py
Added variable to control what info is displayed. 
if showSSID, showPASS, or showMoreNetinfo are False. They will not be displayed, in the case of showMoreNetinfo it will not try to connect to the AP and internet to get its IP. 

QR code is still displayed if everything else is set to false.